chore: rename security-svc to jpo-security-svcs for consistency
