+++ docker-compose.yml
@@ -1,7 +1,7 @@
 version: '3'
 services:
   sec:
-    build: ./security-svc/
+    build: jpo-security-svcs/
     image: jpoode_sec:latest
     ports:
      - "8090:8090"
