+++ jpo-security-svcs/src/main/resources/logback.xml
@@ -1,0 +1,16 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<configuration debug="true" scan="true">
+	<jmxConfigurator />
+
+	<appender name="STDOUT" class="ch.qos.logback.core.ConsoleAppender">
+		<encoder>
+			<pattern>%date{"yyyy-MM-dd HH:mm:ss", UTC} [%thread] %-5level %logger{0} - %msg %n</pattern>
+		</encoder>
+	</appender>
+	
+	<logger name="org.springframework.web" level="INFO" />
+	<logger name="us.dot.its.jpo.sec" level="DEBUG" />
+	<root level="WARN">
+		<appender-ref ref="STDOUT" />
+	</root>
+</configuration>
