+++ jpo-security-svcs/src/test/java/us/dot/its/jpo/sec/models/MessageTest.java
